commit Tom

+++ DV_Projekt_Schnittdatenrechner/src/Drehen.java
@@ -1,6 +1,6 @@
 
 public class Drehen {
 	
-	 int vorschub;
+	 double vorschub;
 
 }
